Placeholder blocks and solved memory leak
So I really didn't get as far as I meant to on optimizing the blocks. I
spent a really long time trying to figure out why everything is
breaking. I realized I wasn't deleting the array in the heap. As it
stands blocks should be functional in the vector of pointers. I'll look
into fixing the blocks appearance more later.